feat(lotto): 콘솔 출력 로직을 ResultView class로 이동
