+++ src/main/java/lotto/LottoGame.java
@@ -4,6 +4,7 @@ import lotto.domain.Lotto;
 import lotto.domain.LottoMachine;
 import lotto.domain.Price;
 import lotto.ui.InputView;
+import lotto.ui.ResultView;
 
 import java.util.List;
 
@@ -11,13 +12,15 @@ public class LottoGame {
 
     public static void main(String[] args) {
         InputView inputView = new InputView();
+        ResultView resultView = new ResultView();
+
         Price price = inputView.insertPrice();
 
         LottoMachine lottoMachine = new LottoMachine();
         List<Lotto> lottoList = lottoMachine.buy(price);
-        inputView.answerPrice(lottoList.size());
+        resultView.printLottoCount(lottoList.size());
 
-        lottoList.forEach(inputView::printLotto);
+        lottoList.forEach(resultView::printLotto);
 
         Lotto winLotto = inputView.insertWinLotto();
     }
